Mariana TQ before meeting 28/02/2018

+++ lib/sampler_20171119.cpp
@@ -21,7 +21,7 @@ void Sampler::initialize(void){
 bool Sampler::runBlock(void) {
 
 	outputSignals[0]->setSamplingPeriod(inputSignals[0]->getSymbolPeriod());
-	numberOfInputSignals = inputSignals.size();
+	numberOfInputSignals = (int)inputSignals.size();
 
 	bool alive{ false };
